+++ src/threads.js
@@ -9,7 +9,7 @@
     written by Jens Mönig
     jens@moenig.org
 
-    Copyright (C) 2025 by Jens Mönig
+    Copyright (C) 2026 by Jens Mönig
 
     This file is part of Snap!.
 
@@ -66,7 +66,7 @@ CustomHatBlockMorph*/
 
 /*jshint esversion: 11, bitwise: false, evil: true*/
 
-modules.threads = '2025-December-19';
+modules.threads = '2026-January-03';
 
 var ThreadManager;
 var Process;
@@ -9457,7 +9457,7 @@ Process.prototype.doSetBlockAttribute = function (attribute, block, val) {
         template.refreshDefaults();
     }
     ide.flushPaletteCache();
-    ide.categories.refreshEmpty();
+    ide.refreshEmptyCategories();
     ide.refreshPalette();
     rcvr.recordUserEdit(
         'scripts',
@@ -9533,7 +9533,7 @@ Process.prototype.doDefineBlock = function (upvar, label, context) {
 
     // update the IDE
     ide.flushPaletteCache();
-    ide.categories.refreshEmpty();
+    ide.refreshEmptyCategories();
     ide.refreshPalette();
     rcvr.recordUserEdit(
         'palette',
@@ -9641,7 +9641,7 @@ Process.prototype.doDeleteBlock = function (context) {
 
     // update the IDE
     ide.flushPaletteCache();
-    ide.categories.refreshEmpty();
+    ide.refreshEmptyCategories();
     ide.refreshPalette();
     rcvr.recordUserEdit(
         'palette',